requirements.txt
fastapi
uvicorn
faster-whisper
edge-tts
motor
dnspython
passlib[bcrypt]
python-multipart
python-dotenv
